restart crashed scripts

+++ index.js
@@ -13,6 +13,9 @@ var child;
 
 function spawn() {
   child = cp.spawn('node', args, { stdio: ['pipe', process.stdout, process.stderr] });
+  child.on('exit', function (c, s) {
+    child = null; // the child process might crash and/or end by itself (not only by our kill)
+  });
 }
 
 var grace_period = 1;
@@ -24,10 +27,16 @@ watcher.on('modify', function () {
   var delta = process.hrtime(last_modify);
   if (delta[0] > grace_period) {
     last_modify = process.hrtime();
-    child.kill();
-    child.on('exit', function (c, s) {
-      process.nextTick(spawn);
-    });
+    if (child) {
+      // child not null, register a spawn and kill it
+      child.on('exit', function (c, s) {
+        spawn();
+      });
+      child.kill();
+    } else {
+      // child is null, just spanw it
+      spawn();
+    }
   }
 });
